Dev (#5)
* Add dev build action

* Access check (#2)

* Add code for checking user's access to a path

* added access checks

* added access checks

* Add fh checks

* added access checks for fh

* fixed include

* added access checks for fh

* added access checks for fh

* added handling for root path

* bugfix

* empty line removed

* fixed check_access for equal paths

* debug

* debug

* debug

* debug

* debug

* debug

* fix when fh is not set

* Use lexically_normal

* debug

* debug

* debug

* debug

* insert fh in more conditions

* insert fh in more conditions

* removed debug


* Run on ubuntu 20

* fix for group folders

* fix for vscode

* fix for vscode

* Add more logs

* Add more logs

* Add more logs

* Add more logs

* Debug libreoffice

* revert release change

* Add more logs

* Use open instead of creat

* Add more logs

* Add more logs

* Fix flag size on create

* Fix create

* Fix create

* Fix create

* fix for vscode/libreoffice

* fix for vscode/libreoffice

---------

+++ capnp/getattr.capnp
@@ -5,7 +5,7 @@ struct Getattr {
   fi    @1 :FuseFileInfo;
 
   struct FuseFileInfo {
-    flags         @0  :Int8;
+    flags         @0  :Int64;
     writepage     @1  :UInt64;
     directIo      @2  :UInt64;
     keepCache     @3  :UInt64;
